Damaged Product refactored

+++ src/IICT-Store.Models/Products/DamagedProductSerialNo.cs
@@ -10,7 +10,7 @@ namespace IICT_Store.Models.Products
     {
         public string Name { get; set; }
         public long ProductNoId { get; set; }
-        public int DamagedProductId { get; set; }
+        public long DamagedProductId { get; set; }
         public DamagedProduct DamagedProduct { get; set; }
     }
 }
